The Bisection Method
In general, if f(x) is real and continuous in the interval from xl to xu and f(xl) and f(xu) have opposite signs, that is, f(xl) *f(xu)  < 0, then there is at least one real root between xl and xu. 

Incremental search methods capitalize on this observation by locating an interval where the function changes sign. Then the location of the sign change (and consequently, 
the root) is identified more precisely by dividing the interval into a number of subintervals. Each of these subintervals is searched to locate the sign change. The process is 
repeated and the root estimate refined by dividing the subintervals into finer increments. 

The bisection method, which is alternatively called binary chopping, interval halving, or Bolzano’s method, is one type of incremental search method in which the interval is 
always divided in half. If a function changes sign over an interval, the function value at the midpoint is evaluated. The location of the root is then determined as lying at the 
midpoint of the subinterval within which the sign change occurs. The process is repeated to obtain refined estimates. 

In this program, I've implemented the Bisection Method to determine the drag coefficient of a free falling parachutist. However, this program can be used to determine the root of any equation simply by modifying the function definition.